refactor(ai-chat): extract formatRelativeDate helper
