+++ plugins/ai-chat/src/components/AiChatDrawer/DrawerConversationHistory.tsx
@@ -20,7 +20,7 @@ import { useState } from 'react';
 import { useConversations } from '../../hooks/useConversations';
 import { DeleteConversationDialog } from '../DeleteConversationDialog';
 import type { ConversationApi, ConversationListItem } from '../../api';
-import { getConversationTitle } from '../../utils';
+import { formatRelativeDate, getConversationTitle } from '../../utils';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -53,21 +53,6 @@ const useStyles = makeStyles(theme => ({
     },
   },
 }));
-
-function formatRelativeDate(dateStr: string): string {
-  const date = new Date(dateStr);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffMins < 1) return 'Just now';
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  if (diffDays < 7) return `${diffDays}d ago`;
-  return date.toLocaleDateString();
-}
 
 interface DrawerConversationHistoryProps {
   conversationApi: ConversationApi;
