+++ plugins/ai-chat/src/utils/index.ts
@@ -1,1 +1,2 @@
+export { formatRelativeDate } from './formatRelativeDate';
 export { getConversationTitle } from './getConversationTitle';
